\name{b1}
\alias{b1}
\alias{b2}
\docType{data}
\title{
mcwBestCis instances,
integrative analysis output containers generated by GGtools vignette
}
\description{
integrative analysis output containers generated by GGtools vignette
}
\usage{data(b1)}
\format{
  The format is: \cr
Formal class 'mcwBestCis' [package "GGtools"] with 9 slots \cr
  ..@ scoregr  :Formal class 'GRanges' [package "GenomicRanges"] with 6 slots \cr
  .. .. ..@ seqnames       :Formal class 'Rle' [package "IRanges"] with 4 slots \cr
  .. .. .. .. ..@ values         : Factor w/ 1 level "20": 1 \cr
  .. .. .. .. ..@ lengths        : int 50 \cr
  .. .. .. .. ..@ elementMetadata: NULL \cr
  .. .. .. .. ..@ metadata       : list() \cr
  .. .. ..@ ranges         :Formal class 'IRanges' [package "IRanges"] with 6 slots \cr
  .. .. .. .. ..@ start          : int [1:50] 24280834 61665697 352356 61679079 45286150 55187941 38766161 10871477 56570242 13304639 ... \cr
  .. .. .. .. ..@ width          : int [1:50] 2090785 2005619 2021461 2001901 2129211 2007692 2038197 2035767 2012068 2013675 ... \cr
  .. .. .. .. ..@ NAMES          : chr [1:50] "GI_34147330-S" "hmm26961-S" "GI_17149835-I" "GI_31077201-S" ... \cr
  .. .. .. .. ..@ elementType    : chr "integer" \cr
  .. .. .. .. ..@ elementMetadata: NULL \cr
  .. .. .. .. ..@ metadata       : list() \cr
  .. .. ..@ strand         :Formal class 'Rle' [package "IRanges"] with 4 slots \cr
  .. .. .. .. ..@ values         : Factor w/ 3 levels "+","-","*": 3 \cr
  .. .. .. .. ..@ lengths        : int 50 \cr
  .. .. .. .. ..@ elementMetadata: NULL \cr
  .. .. .. .. ..@ metadata       : list() \cr
  .. .. ..@ elementMetadata:Formal class 'DataFrame' [package "IRanges"] with 6 slots \cr
  .. .. .. .. ..@ rownames       : NULL \cr
  .. .. .. .. ..@ nrows          : int 50 \cr
  .. .. .. .. ..@ listData       :List of 6 \cr
  .. .. .. .. .. ..$ score     : num [1:50] 36.5 16.8 16.7 14.8 9.8 ... \cr
  .. .. .. .. .. ..$ snpid     : chr [1:50] "rs6037097" "rs3810504" "rs13043344" "rs13044229" ... \cr
  .. .. .. .. .. ..$ snploc    : int [1:50] 25347221 62678549 544417 61738288 46369894 56551001 38879237 12177765 57565943 15056960 ... \cr
  .. .. .. .. .. ..$ radiusUsed: num [1:50] 1e+06 1e+06 1e+06 1e+06 1e+06 1e+06 1e+06 1e+06 1e+06 1e+06 ... \cr
  .. .. .. .. .. ..$ nsnp      : int [1:50] 1387 783 2382 766 2157 2238 1658 2459 1675 1868 ... \cr
  .. .. .. .. .. ..$ fdr       : num [1:50] 0 0 0 0.25 0.717 ... \cr
  .. .. .. .. ..@ elementType    : chr "ANY" \cr
  .. .. .. .. ..@ elementMetadata: NULL \cr
  .. .. .. .. ..@ metadata       : list() \cr
  .. .. ..@ seqinfo        :Formal class 'Seqinfo' [package "GenomicRanges"] with 4 slots \cr
  .. .. .. .. ..@ seqnames   : chr "20" \cr
  .. .. .. .. ..@ seqlengths : int NA \cr
  .. .. .. .. ..@ is_circular: logi NA \cr
  .. .. .. .. ..@ genome     : chr NA \cr
  .. .. ..@ metadata       : list() \cr
  ..@ allperm  : num [1:100] 15.8 13.9 13.8 13.2 13 ... \cr
  ..@ extra    : NULL \cr
  ..@ chromUsed: chr "20" \cr
  ..@ theCall  : language best.cis.eQTLs(smpack = "GGdata", rhs = ~male, folderstem = "db2", radius = 1e+06,      chrnames = "20", geneApply = mclapply, snpannopk = snplocsDefault(),  ... \cr
  ..@ smFilter :function (x)   \cr
  .. ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 413 8 413 50 8 50 4 4 \cr
  .. .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0xf5c0cc0>  \cr
  ..@ nperm    : num 2 \cr
  ..@ globalMap:<environment: 0xf5c0538>  \cr
  ..@ testCount: int 83969 \cr
}
\details{
As created in GGtools.Rnw vignette code, with sharply curtailed searches
}
%\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%%%  ~~ possibly secondary sources and usages ~~
%}
\examples{
data(b1)
b1
}
\keyword{datasets}
