% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_annual_emissions.R
\name{chart_annual_emissions_by}
\alias{chart_annual_emissions_by}
\alias{chart_annual_emissions}
\title{chart_annual_emissions_by}
\usage{
chart_annual_emissions_by(annual_emission_data, ..., geom = NULL,
  flag_years = NULL, year_limits = c(1990, 2035), title = NULL,
  subtitle = NULL, caption = NULL, verbose = TRUE)

chart_annual_emissions(annual_emission_data, ..., geom = NULL,
  flag_years = NULL, year_limits = c(1990, 2035), title = NULL,
  subtitle = NULL, caption = NULL, verbose = TRUE)
}
\arguments{
\item{annual_emission_data}{tabular, with at least \code{year}, \code{pol_abbr}, \code{ems_qty}, and \code{ems_unit}}

\item{...}{not used}

\item{geom}{supply "line" or "point" if you don't want filled areas}

\item{flag_years}{if provided, will display numeric values above these}

\item{title}{passed to \code{labs()}}

\item{subtitle}{passed to \code{labs()}}

\item{caption}{passed to \code{labs()}}

\item{verbose}{display messages}
}
\description{
Experimental support for easy charting of annual emissions.
}
\note{
\strong{Please don't use in production code! Thank you!}
}
