% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_annual_emissions.R
\name{chart_annual_emissions_by}
\alias{chart_annual_emissions_by}
\title{chart_annual_emissions_by}
\usage{
chart_annual_emissions_by(annual_emission_data, ..., geom = NULL,
  flag_years = NULL, title = NULL, subtitle = NULL, caption = NULL,
  verbose = TRUE)
}
\description{
chart_annual_emissions_by
}
