% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_annual_quantities_by.R
\name{chart_annual_quantities}
\alias{chart_annual_quantities}
\title{chart_annual_quantities}
\usage{
chart_annual_quantities(
  data = NULL,
  ...,
  mapping = aes(),
  geom = NULL,
  qty_var = NULL,
  chart_y_scale = NULL,
  chart_y_labels = NULL,
  chart_y_title = waiver(),
  chart_y_unit = NULL,
  chart_gridlines = TRUE,
  chart_geom = NULL,
  facet_rows = NULL,
  facet_cols = NULL,
  facet_scales = "fixed",
  year_limits = NULL,
  year_breaks = NULL,
  year_expand = NULL,
  flag_years = NULL,
  flag_labels = NULL,
  flag_unique = NULL,
  base_year = NULL,
  base_year_shape = 3,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  verbose = getOption("verbose")
)
}
\description{
chart_annual_quantities
}
