% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_by_pollutant.R
\name{facet_by_pollutant}
\alias{facet_by_pollutant}
\title{facet_by_pollutant}
\usage{
facet_by_pollutant(
  chart_object,
  facet_var = "pol_abbr",
  verbose = getOption("verbose")
)
}
\description{
facet_by_pollutant
}
\note{
used by \code{\link[=chart_annual_emissions_by]{chart_annual_emissions_by()}} and \code{\link[=chart_annual_emission_factors_by]{chart_annual_emission_factors_by()}}
}
