% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_swatches.R
\name{fct_swatches}
\alias{fct_swatches}
\title{fct_swatches}
\usage{
fct_swatches(x, fun = scale_color_tableau, ...)
}
\arguments{
\item{x}{(factor) or character, convertible to factor}

\item{fun}{(function) like \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_color_discrete()}}}

\item{...}{further arguments to \code{fun}}
}
\description{
Generate colors for a given factor, or something coercible to one.
Useful in conjunction with \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.
}
