% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rate.R
\name{get_rate_with_cusum}
\alias{get_rate_with_cusum}
\title{cumsum(累積和)したときに占める割合を求める}
\usage{
get_rate_with_cusum(x, arrange = TRUE)
}
\arguments{
\item{x}{numeric 累積和するnumeric型のベクトル}

\item{arrange}{boolean TRUEなら昇順にしてから求める.FALSEならそのまま}
}
\value{
numeric 全体の何割を占めているかのベクトル
}
\description{
cumsum(累積和)したときに占める割合を求める
}
