% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_emission_chart.R
\name{gg_emission_chart}
\alias{gg_emission_chart}
\title{Plot annual emissions}
\usage{
gg_emission_chart(chart_data, limits = list(x = NULL, y = c(0, NA)),
  caption = "DRAFT {format(Sys.Date(), '\%Y-\%m-\%d')}",
  verbose = TRUE)
}
\description{
Plot annual emissions
}
\seealso{
\itemize{
\item \code{\link[=gg_throughput_chart]{gg_throughput_chart()}}
}
}
