% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_throughput_chart.R
\name{gg_throughput_chart}
\alias{gg_throughput_chart}
\title{Use gg_chart() to plot tput_qty}
\usage{
gg_throughput_chart(chart_data, ...)
}
\description{
Use gg_chart() to plot tput_qty
}
\note{
group your data first to alter its presentation
}
