% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_piechart.R
\name{ggplot2_piechart}
\alias{ggplot2_piechart}
\title{ggplot2-style piechart}
\usage{
ggplot2_piechart(
  input_data,
  ...,
  colors = NULL,
  digits = 0,
  sorted = TRUE,
  theme = NULL
)
}
\arguments{
\item{input_data}{data frame}

\item{...}{further arguments (optional)}

\item{colors}{named character}

\item{digits}{digits after decimal}

\item{sorted}{logical}
}
\description{
ggplot2-style piechart
}
\examples{
BY2011_NOx_2015 <- BY2011_annual \%>\% filter(year == 2015, pol_abbr == "NOx")
BY2011_NOx_2015_by_cnty <- BY2011_NOx \%>\% annual_emissions_by(cnty_abbr)
BY2011_NOx_2015_by_cnty \%>\% select(cnty_abbr, ems_qty) \%>\% arrange(ems_qty) \%>\% ggplot2_piechart()
}
