% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtools-package.R
\docType{package}
\name{ggtools}
\alias{ggtools}
\alias{ggtools-package}
\title{ggtools: GGplot Tools}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: t k \email{t.k@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{GGPlot}
\keyword{Tools}
\keyword{internal}
