% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_log10.R
\name{label_log10}
\alias{label_log10}
\title{label_log10}
\usage{
label_log10(x, digits = 1, TeX = TRUE, output = "expression")
}
\arguments{
\item{x}{numeric vector}

\item{digits}{integer}

\item{TeX}{logical}

\item{output}{passed to \code{\link[latex2exp:TeX]{latex2exp::TeX()}}}
}
\description{
label_log10
}
