% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart.R
\name{make_piechart_text_position}
\alias{make_piechart_text_position}
\title{Caluculating the text posion in pie chart}
\usage{
make_piechart_text_position(y)
}
\arguments{
\item{y}{numeric pie chartで描く値}
}
\value{
numeric textのポジション
}
\description{
Caluculating the text posion in pie chart
}
\examples{
make_piechart_text_position(vector)

}
