% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2palette.R
\name{scale2palette}
\alias{scale2palette}
\title{scale2palette}
\usage{
scale2palette(scale)
}
\arguments{
\item{scale}{a ggplot2 scale object}
}
\value{
character
}
\description{
Extracts color values from a ggplot2 scale
}
\note{
Used by \code{\link[=scale_sepia_fill]{scale_sepia_fill()}} and \code{\link[=scale_sepia_color]{scale_sepia_color()}}
}
