% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pollutant.R
\name{scale_color_pollutant}
\alias{scale_color_pollutant}
\title{scale_color_pollutant}
\usage{
scale_color_pollutant(...)
}
\description{
Proposed convention for coloring emissions by pollutant.
}
\note{
FIXME: implement this
}
