% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_cubehelix.R
\name{scale_cubehelix}
\alias{scale_cubehelix}
\alias{scale_fill_cubehelix}
\alias{scale_color_cubehelix}
\title{scale_cubehelix}
\usage{
scale_cubehelix(name, aesthetics, ..., start = 0, rot = 0.4,
  gamma = 1, chrom = 0.8, light = 0.85, dark = 0.15)

scale_fill_cubehelix(...)

scale_color_cubehelix(...)
}
\arguments{
\item{name}{as in \code{ggplot2::continuous_scale()}}

\item{...}{further arguments to \code{ggplot2::continuous_scale()}}

\item{start}{see Details}

\item{rot}{see Details}

\item{gamma}{see Details}

\item{light}{see Details}

\item{dark}{see Details}
}
\description{
scale_cubehelix

scale_fill_cubehelix

scale_color_cubehelix
}
\details{
starting hue (here, \code{start}) is:
\itemize{
\item \code{start} in \code{seaborn.cubehelix_palette}
\item \code{h} in \code{chroma::cubehelix_map()}
}

chromaticity is:
\itemize{
\item \code{hue} in \code{seaborn.cubehelix_palette}
\item \code{c} in \code{chroma::cubehelix_map()}
}

lightness and darkness:
\itemize{
\item valid range is \link{0, 1} in both.
}

gamma:
\itemize{
\item valid range is [0, Inf) in both
\item default is 1.0
\item sane values seem to be in \link{0.5, 3} (whitened, darkened)
}
}
\seealso{
\itemize{
\item \code{chroma::cubehelix_scale()}
\item \code{chroma::cubehelix_map()}
}
}
