% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_extent.R
\name{scale_extent}
\alias{scale_extent}
\title{scale_extent}
\usage{
scale_extent(object, ..., verbose = getOption("verbose", default = TRUE))
}
\arguments{
\item{object}{\code{sf}, \link[raster:Extent-class]{raster::Extent}, or \link[terra:SpatExtent-class]{terra::SpatExtent} object}

\item{...}{passed to \code{\link[=scale_easting]{scale_easting()}} and \code{\link[=scale_northing]{scale_northing()}}}

\item{verbose}{(logical)}
}
\description{
scale_extent
}
