% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_cubehelix.R
\name{scale_fill_cubehelix}
\alias{scale_fill_cubehelix}
\title{scale_fill_cubehelix}
\usage{
scale_fill_cubehelix(..., start = 0, rot = 0.4, gamma = 1,
  chrom = 0.8, light = 0.85, dark = 0.15, guide = "colorbar")
}
\description{
scale_fill_cubehelix
}
\details{
starting hue (here, \code{start}) is:
\itemize{
\item \code{start} in \code{seaborn.cubehelix_palette}
\item \code{h} in \code{chroma::cubehelix_map}
}

chromaticity is:
\itemize{
\item \code{hue} in \code{seaborn.cubehelix_palette}
\item \code{c} in \code{chroma::cubehelix_map} (chromaticity)
}

lightness and darkness:
\itemize{
\item valid range is \link{0, 1} in both.
}

gamma:
\itemize{
\item valid range is [0, Inf) in both
\item default is 1.0
\item sane values seem to be in \link{0.5, 3} (whitened, darkened)
}
}
