% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_log10.R
\name{scale_x_log10}
\alias{scale_x_log10}
\alias{scale_y_log10}
\title{scale_x_log10}
\usage{
scale_x_log10(...)

scale_y_log10(...)
}
\description{
scale_x_log10

scale_y_log10
}
\section{Functions}{
\itemize{
\item \code{scale_x_log10}: x scale

\item \code{scale_y_log10}: y scale
}}

