% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_quantity.R
\name{scale_quantity}
\alias{scale_quantity}
\alias{scale_x_quantity}
\alias{scale_y_quantity}
\alias{scale_xy_quantity}
\title{scale_quantity}
\usage{
scale_quantity(
  name = NULL,
  dimension = c("x", "y"),
  expand = ggplot2::expansion(mult = c(0, 0.1)),
  limits = c(0, NA),
  unit = NULL,
  labels = NULL,
  format = NULL,
  ...,
  verbose = getOption("verbose")
)

scale_x_quantity(...)

scale_y_quantity(...)

scale_xy_quantity(...)
}
\arguments{
\item{name}{(character) axis label}

\item{expand}{passed to \code{scale_continuous()}}

\item{limits}{passed to \code{scale_continuous()}}

\item{unit}{(character) passed to \code{format_qty()}}

\item{labels}{passed to \code{scale_continuous()}}

\item{format}{passed to \code{scale_continuous()}}

\item{...}{passed to \code{scale_continuous()}}

\item{verbose}{(logical) display messages}
}
\description{
Position scales for quantities (x & y)
}
\section{Functions}{
\itemize{
\item \code{scale_x_quantity()}: 

\item \code{scale_y_quantity()}: 

\item \code{scale_xy_quantity()}: 

}}
