% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_sepia.R
\name{scale_sepia}
\alias{scale_sepia}
\alias{scale_fill_sepia}
\alias{scale_color_sepia}
\title{Continuous scales for air quality}
\usage{
scale_sepia(
  ...,
  guide = "colorbar",
  start = 48,
  rot = -0.14,
  chrom = 0.95,
  light = 0.97,
  dark = 0.1,
  gamma = 0.8,
  na.value = "#00000000",
  aesthetics = c("fill", "color")
)
}
\arguments{
\item{...}{passed to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}}

\item{guide}{as in \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}}

\item{start}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{rot}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{chrom}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{light}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{dark}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{gamma}{as in \code{\link[chroma:cubehelix_scale]{chroma::cubehelix_scale()}}}

\item{na.value}{color to use for \code{NA} values (default transparent, but may not work)}

\item{aesthetics}{for internal use}
}
\description{
Built to resemble ArcGIS-style "pollution" color ramps.
}
