% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_annual.R
\name{scale_x_annual}
\alias{scale_x_annual}
\title{Annual scale for CY, RY, PY, etc.}
\usage{
scale_x_annual(name = NULL, prefix = "", breaks = seq(1990, 2050, by
  = 5), labels = function(x) str_extract(x, "[0-9]{4}"),
  limits = waiver(), expand = expand_scale(add = c(0.2, 0.2)), ...,
  verbose = getOption("verbose"))
}
\description{
Annual scale for CY, RY, PY, etc.
}
