% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_decades.R
\name{scale_x_decades}
\alias{scale_x_decades}
\title{Scale for Years}
\usage{
scale_x_decades(
  name = "",
  ...,
  limits = c(1990, 2050),
  breaks = seq(1990, 2050, by = 20)
)
}
\arguments{
\item{name}{(character) axis label}

\item{limits}{(numeric) axis range}

\item{breaks}{(numeric) see \code{\link[ggplot2:scale_x_continuous]{ggplot2::scale_x_continuous()}}}
}
\description{
Scale for Years
}
