% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_ecdf.R
\name{scale_y_ecdf}
\alias{scale_y_ecdf}
\title{scale_y_ecdf}
\usage{
scale_y_ecdf(
  name = "Cumulative distribution",
  ...,
  trans = "identity",
  digits = 1,
  limits = c(0, 1),
  breaks = waiver(),
  expand = expansion(0, 0)
)
}
\arguments{
\item{...}{(numeric) passed to \code{\link[=scale_y_percentage]{scale_y_percentage()}}}

\item{trans}{(character) useful choices are "logit" and "identity" (the default).}

\item{digits}{(numeric) passed to \code{\link[=scale_y_percentage]{scale_y_percentage()}}}

\item{limits}{(numeric) passed to \code{\link[=scale_y_percentage]{scale_y_percentage()}}}

\item{breaks}{(numeric) passed to \code{\link[=scale_y_percentage]{scale_y_percentage()}}}

\item{expand}{(numeric) passed to \code{\link[=scale_y_percentage]{scale_y_percentage()}}}
}
\description{
scale_y_ecdf
}
