% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_emissions.R
\name{scale_y_emissions}
\alias{scale_y_emissions}
\title{scale_y_emissions}
\usage{
scale_y_emissions(
  name = c("tons/yr", "tons/day"),
  labels = NULL,
  expand = expand_scale(mult = c(0, 0.1)),
  limits = c(0, NA),
  ...,
  verbose = getOption("verbose")
)
}
\description{
scale_y_emissions
}
