% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_percentage.R
\name{scale_y_percentage}
\alias{scale_y_percentage}
\title{scale_y_percentage}
\usage{
scale_y_percentage(..., limits = c(0, 1), breaks = seq(-100, 100, by =
  0.2), labels = format_percentage, expand = expand_scale(mult = 0, add
  = 0))
}
\description{
scale_y_percentage
}
