% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_SI_units.R
\name{scale_y_SI_units}
\alias{scale_y_SI_units}
\title{Scale for Quantities}
\usage{
scale_y_SI_units(..., limits = c(0, NA), labels = format_SI)
}
\arguments{
\item{...}{passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}

\item{limits}{(numeric) axis range}

\item{labels}{see \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}}
}
\description{
So that every chart can have a simple Y scale fixed at 0 and with SI units.
}
