% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_y_throughputs.R
\name{scale_y_throughputs}
\alias{scale_y_throughputs}
\title{scale_y_throughputs}
\usage{
scale_y_throughputs(
  name,
  labels = NULL,
  expand = ggplot2::expansion(mult = c(0, 0.1)),
  limits = c(0, NA),
  ...,
  verbose = getOption("verbose")
)
}
\description{
scale_y_throughputs
}
