\name{snplocsDefault}
\alias{snplocsDefault}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
name the default SNPlocs.Hsapiens.dbSNP.* package
}
\description{
generate a string naming the default SNPlocs.Hsapiens.dbSNP.* package
for use with GGtools
}
\usage{
snplocsDefault()
}
%- maybe also 'usage' for other objects documented here.
\details{
allows centralized specification of SNPlocs resource package
}
\value{
a character string, see example
}
%\references{
%% ~put references to the literature/web site here ~
%}
%\author{
%%%  ~~who you are~~
%}
%\note{
%%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
\examples{
snplocsDefault()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ models }
