% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gridlines.R
\name{theme_gridlines}
\alias{theme_gridlines}
\title{theme_gridlines}
\usage{
theme_gridlines(linetype = "dotted", color = gray(0.9), ...)
}
\arguments{
\item{linetype}{(character) passed to \code{\link[ggplot2:element]{ggplot2::element_line()}}}

\item{color}{(character) passed to \code{\link[ggplot2:element]{ggplot2::element_line()}}}

\item{...}{further arguments to \code{\link[ggplot2:element]{ggplot2::element_line()}}}
}
\description{
Subtle gridlines for a ggplot object.
Useful in conjunction with \code{\link[=theme_simple]{theme_simple()}}.
}
