% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_remove.R
\name{theme_remove}
\alias{theme_remove}
\title{theme_remove}
\usage{
theme_remove(...)
}
\arguments{
\item{...}{(character) see Details and examples}
}
\description{
Removes specified theme elements by setting them to \code{\link[ggplot2:element]{ggplot2::element_blank()}}.
}
\details{
All combinations of the elements in \code{...} will be removed.
See examples.
}
\examples{
p <- ggplot(mtcars) + geom_point()
p + theme_remove("axis", "text")
p + theme_remove("axis", c("text", "line"), "x")
p + theme_remove("axis", c("text", "line"), c("x", "y"))
}
