\name{GGtools-trackSet-methods}
\docType{methods}
\alias{trackSet-methods}
\alias{trackSet,cwSnpScreenResult-method}
\title{ create a browser track from a chromosome-wide SNP screen}
\description{ create a browser track from a chromosome-wide SNP screen}
\section{Methods}{
\describe{


\item{object = "cwSnpScreenResult"}{ returns a track set with genomic
coordinates for x axis and -log10 p-values for y axis}

}}
\keyword{methods}
