% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_annual_quantities.R
\name{chart_annual_quantities}
\alias{chart_annual_quantities}
\title{chart_annual_quantities}
\usage{
chart_annual_quantities(chart_data, year_limits = c(NA, NA),
  year_breaks = seq(1990, 2050, by = 10),
  year_expand = expand_scale(add = c(1, 1), mult = c(0, 0)), ...)
}
\description{
chart_annual_quantities
}
\note{
foundation for chart_annual_*
}
\seealso{
\itemize{
\item \code{\link[=chart_annual_growth]{chart_annual_growth()}}
\item \code{\link[=chart_annual_emissions]{chart_annual_emissions()}}
}
}
