% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_swatches.R
\name{fct_swatches}
\alias{fct_swatches}
\title{fct_swatches}
\usage{
fct_swatches(x, fun = scale_color_tableau, ...)
}
\arguments{
\item{x}{(factor) or character, convertible to factor}

\item{fun}{(function) like \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_color_discrete()}}}

\item{...}{further arguments to \code{fun}}
}
\value{
named vector, with colors as values, and levels as names
}
\description{
Generate colors for a given factor, or something coercible to one.
Useful in conjunction with \code{\link[ggplot2:scale_manual]{ggplot2::scale_color_manual()}}.
}
\examples{
df <- mutate(mtcars, carb = fct_other(factor(carb), keep = 1:2))
p <- ggplot(df, aes(wt, mpg)) + geom_point(aes(color = carb))
swatches <- replace(fct_swatches(p$data$carb), "Other", gray(0.5))
p + scale_color_manual(values = swatches)
}
