% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtools-package.R
\docType{package}
\name{ggtools}
\alias{ggtools}
\alias{ggtools-package}
\title{ggtools: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{GGPlot}
\keyword{Tools}
\keyword{internal}
