% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PieChart.R
\name{make_piechart_text_position}
\alias{make_piechart_text_position}
\title{pie chartのテキストのポジションを求める}
\usage{
make_piechart_text_position(y)
}
\arguments{
\item{y}{numeric pie chartで描く値}
}
\value{
numeric textのポジション
}
\description{
pie chartのテキストのポジションを求める
}
\examples{
make_piechart_text_position(c(30, 10, 20))
}
