% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_log10.R
\name{scale_x_log10}
\alias{scale_x_log10}
\title{log10 scale together with axis ticks}
\usage{
scale_x_log10(..., labels = format_log10)
}
\arguments{
\item{...}{passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_log10()}}}

\item{labels}{function}
}
\description{
log10 scale together with axis ticks
}
\seealso{
\code{\link[=scale_xy_log10]{scale_xy_log10()}}
}
